s6-rc-hook: check if service exists before restart
If the user doesn't have the service installed, then the longrun_restart
call will fail. Not that this does any harm, but just check to see if we
have the directory before trying to restart it. Additionally, the
directory name may be in the form of ${sv}-srv so check that as well and
restart if it does exist (like for atd for example).