s6: kill unwanted supervisors in db hook
The s6-rc-update logic has a bit of a quirk to it. Longruns that are
down will not have their s6-supervise process killed when the database
is upgraded. This means that after a lot of database upgrades a user
could have their ps output littered with a ton of useless s6-supervise
processes. Previously, the hook killed such processes when a script was
being removed but we can do better. Instead, kill the s6-supervise
process for all down longruns right before we upgrade the database. If a
longrun is up, the s6-supervise process survives and and carries on to
the database upgrade so no need to worry there. This approach should
keep the ps output much cleaner and avoid bloating the user with a ton
of annoying (albiet harmless since they just do nothing) s6-supervise
processes.