s6-rc-db-update-hook: make database update atomic
Technically, this has been wrong for a while (no bad side effects seem
to have been reported at least). The original version of this script did
do atomic updates but it got lost in a recent-ish rewrite. According to
skarnet documentation, the updated symlink to the newly compiled
database should be made atomically. The standard ln utilty does not do
that and actually forbids this. We could use s6-ln (it is packaged), but
artix is a GNU distro so it's better not to incur any additional
dependencies. The correct way to fix this is to simply use && mv -f as
per the skarnet documentation.