s6-rc-db-update: optimize the main update loop
s6-rc-bundle-update must do at least one loop to update an existing
bundle. In the main update loop, s6-rc-bundle-update is called several
times which means we do several loops here. However, s6-rc-bundle-update
is designed to be able to handle any arbitrary number of arguments as
needed. In this case, just call it once after we finishing populating
new_contents. This should save a massive amount of time when executing
the script.