s6-rc-db-update-hook: support custom user sv path
Finally got around to working on this. Some users in #artix and #s6
correctly pointed out that Artix's s6 implementation is heavily
dependent on s6-rc-bundle and is not very friendly about letting users
edit and make their own services. Specifically, the default bundle and
its contents are commonly overwritten on every s6-scripts update. This
adds some groundwork to splitting the s6-rc database compilation into
multiple directories. ${USERSVPATH} is the path for custom user service
files. The default bundle will live here along with anything else not
installed by the user. ${BACKUPSVPATH} is strictly for emergency usage
when the compilation fails. These are not to be altered by users.