s6-rc-hook: fix service shutdown behavior
It turns out that old logic caused all services and bundles to always
shutdown. The s6-rc -a list grep always returns some nonzero string.
Thus, the conditional is always true and we always try to shutdown
anything that's removed. This has the side effect of causing any moved
bundles to automatically shutdown as well which is never desirable (only
services that go away for good should shut down). Instead of using -n,
we should check for ! $active which is equivalent to checking if $active
= "". This is what is actually returned by that grep if the sv name is
not in the s6-rc -a list.