s6: update to the new s6-rc format
s6-scripts finally updated to the new s6-rc format so the live-service
needs to change its formating to match. The main advantage of this is
that there is no need to sed in artix-live and pacman-init into the
contents file during the alpm hook. Instead, the empty files can just be
installed into the directory. It's much cleaner. As an aside, a couple
of BMODE calls in the makefile were changed to FMODE since s6-rc doesn't
need any of these files to be executable.